chore: Remove unused test files and update dependencies
This commit removes the unused test files `setupTests.js` and `App.test.js` from the repository. These files were no longer needed and were causing confusion for developers.

Additionally, the commit updates the dependencies in the `package.json` file for both the backend and frontend. It adds the following dependencies:
- `@emotion/react` version 11.11.4
- `@emotion/styled` version 11.11.5
- `@mui/icons-material` version 5.15.21
- `@mui/material` version 5.15.21
- `youtube-dl-exec` version 3.0.6

It also adds the `ajv` dependency as a devDependency in the `package.json` file.

Please review and merge this commit.